Support Scala-Native 0.5.0, bump minimum versions of Scala 2 and Scala 3 accordingly (#73)
This would be the first step in getting the rest of com.lihaoyi out for
scala native 0.5.0

Given the version support drops, which version should we bump here?
Would it be the major version (i.e. geny 2.0.0), or a minor version
(i.e. geny 1.1.0)? I am inclined to make it a minor version, reflecting
the fact that Scala 3.1.3 is not LTS so dropping support is not a huge
deal